feat: cake price
