+++ packages/uikit/src/widgets/SimpleMenu/components/index.ts
@@ -1,4 +1,4 @@
 export * from "./AnnouncementBanner";
-// export * from "./CakePrice";
+export * from "./CakePriceWidget";
 export * from "./LocaleSelector";
 export * from "./Navigation";
